# backend/src/main/java/com/pi1/Edook/dto/FuncionarioResponseDto.java
package com.pi1.Edook.dto;

import lombok.Setter;

@Setter
public class FuncionarioResponseDto {
    private String nome;
    private String cpf;
    private String email;
    private String cargo;
    private Integer matricula;
}