# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/funcionarios/confirmar-email":{"put":{"tags":["funcionario-controller"],"operationId":"confirmarEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReenviarConfirmacaoDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/equipamentos/{prefixo}/{numero}":{"put":{"tags":["equipamento-controller"],"operationId":"atualizar","parameters":[{"name":"prefixo","in":"path","required":true,"schema":{"type":"string"}},{"name":"numero","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipamentoUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EquipamentoResponseDto"}}}}}},"delete":{"tags":["equipamento-controller"],"operationId":"excluir","parameters":[{"name":"prefixo","in":"path","required":true,"schema":{"type":"string"}},{"name":"numero","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/reservas":{"get":{"tags":["reserva-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservaResponseDto"}}}}}}},"post":{"tags":["reserva-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservaCreateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservaResponseDto"}}}}}}},"/funcionarios":{"post":{"tags":["funcionario-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FuncionarioCreateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FuncionarioResponseDto"}}}}}}},"/funcionarios/reenviar-confirmacao":{"post":{"tags":["funcionario-controller"],"operationId":"reenviarConfirmacao","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReenviarConfirmacaoDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/equipamentos":{"get":{"tags":["equipamento-controller"],"operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EquipamentoResponseDto"}}}}}}},"post":{"tags":["equipamento-controller"],"operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipamentoCreateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EquipamentoResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["login-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}}}},"/reservas/{id}/cancelar":{"patch":{"tags":["reserva-controller"],"operationId":"cancelar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"cpf","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ReenviarConfirmacaoDto":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"codigo":{"type":"string","minLength":1}},"required":["codigo","email"]},"EquipamentoUpdateDto":{"type":"object","properties":{"descricao":{"type":"string","minLength":1},"tipo":{"type":"string","minLength":1}},"required":["descricao","tipo"]},"EquipamentoResponseDto":{"type":"object","properties":{"prefixo":{"type":"string"},"numero":{"type":"integer","format":"int32"},"descricao":{"type":"string"},"tipo":{"type":"string"}}},"EquipamentoReservaDto":{"type":"object","properties":{"prefixo":{"type":"string","minLength":1},"numero":{"type":"integer","format":"int32"}},"required":["numero","prefixo"]},"ReservaCreateDto":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"localidade":{"type":"string","minLength":1},"dia":{"type":"string","format":"date"},"horarioInicio":{"type":"string"},"horarioFim":{"type":"string"},"cpfFuncionario":{"type":"string","minLength":1},"equipamentos":{"type":"array","items":{"$ref":"#/components/schemas/EquipamentoReservaDto"},"minItems":1}},"required":["cpfFuncionario","dia","equipamentos","horarioFim","horarioInicio","localidade","nome"]},"ReservaResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nome":{"type":"string"},"localidade":{"type":"string"},"dia":{"type":"string","format":"date"},"horarioInicio":{"type":"string"},"horarioFim":{"type":"string"},"status":{"type":"string"},"cpfFuncionario":{"type":"string"},"equipamentos":{"type":"array","items":{"$ref":"#/components/schemas/EquipamentoResponseDto"}}}},"FuncionarioCreateDto":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"cpf":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"senha":{"type":"string","minLength":1,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&.#_\\-])[A-Za-z\\d@$!%*?&.#_\\-]{6,}$"},"ddd":{"type":"string","minLength":1,"pattern":"\\d{2}"},"numero":{"type":"string","minLength":1,"pattern":"\\d{8,9}"},"cargo":{"type":"string","minLength":1},"matricula":{"type":"integer","format":"int32"},"codigoVerificacao":{"type":"string","minLength":1}},"required":["cargo","codigoVerificacao","cpf","ddd","email","matricula","nome","numero","senha"]},"FuncionarioResponseDto":{"type":"object","properties":{"nome":{"type":"string","writeOnly":true},"cpf":{"type":"string","writeOnly":true},"email":{"type":"string","writeOnly":true},"cargo":{"type":"string","writeOnly":true},"matricula":{"type":"integer","format":"int32","writeOnly":true}}},"EquipamentoCreateDto":{"type":"object","properties":{"prefixo":{"type":"string"},"descricao":{"type":"string"},"tipo":{"type":"string"},"cpfCadastro":{"type":"string"}}},"LoginDto":{"type":"object","properties":{"identificador":{"type":"string","minLength":1},"senha":{"type":"string","minLength":1}},"required":["identificador","senha"]},"LoginResponseDto":{"type":"object","properties":{"nome":{"type":"string"},"cpf":{"type":"string"},"cargo":{"type":"string"},"token":{"type":"string"}}}}}}